Extract AlertContext family + EmailTemplateBuilder to shared lib (Plan E Stage E3a)
AlertContext/AlertDetailItem/AlertContextDto/AlertContextSerializer +
EmailTemplateBuilder move to PerformanceMonitor.Notifications; per-app
copies deleted. Edition string + snooze hint parameterized via a shared
AlertBranding record (Lite passes the snooze hint, Dashboard passes null
-> omits it). Absorbs the PR (b) AlertContext/DTO duplication.
Behaviour-preserving -- emails byte-identical per app (render invariant
test). Gate for E3c.
Plan: ~/.claude/plans/plan-e-alert-service-extraction.md 7.1.

Deviations:
- Lib csproj already had InternalsVisibleTo for both apps + both test
  projects (per plan section 1); no IVT change needed.
- Most consumers already had `using PerformanceMonitor.Notifications;`
  from E1; only 4 files needed it added.
- AlertsHistoryContent.xaml.cs (named in the plan) is NOT a consumer at
  dev@3239790 -- it references none of the moved types.
- Added an internal `EmailAlertService.Branding` test seam (internal,
  test-only via IVT; no public surface) so the render-invariant test
  gates each app's actually-wired branding rather than a re-typed copy.
