--- Lite/Services/AlertContext.cs
@@ -1,114 +1,0 @@
-/*
- * Copyright (c) 2026 Erik Darling, Darling Data LLC
- *
- * This file is part of the SQL Server Performance Monitor Lite.
- *
- * Licensed under the MIT License. See LICENSE file in the project root for full license information.
- */
-
-using System.Collections.Generic;
-using System.Text.Json;
-
-namespace PerformanceMonitorLite.Services;
-
-/// <summary>
-/// Optional detail context attached to alert emails.
-/// Populated from blocking/deadlock detail queries at alert time.
-/// </summary>
-public class AlertContext
-{
-    public List<AlertDetailItem> Details { get; set; } = new();
-    public string? AttachmentXml { get; set; }
-    public string? AttachmentFileName { get; set; }
-}
-
-/// <summary>
-/// A single detail item (e.g., one blocking chain or one deadlock participant).
-/// </summary>
-public class AlertDetailItem
-{
-    public string Heading { get; set; } = "";
-    public List<(string Label, string Value)> Fields { get; set; } = new();
-
-    /// <summary>
-    /// Multi-paragraph prose for this item (advice Investigation / Remediation).
-    /// When non-null, renderers emit this as flowing paragraph text rather than
-    /// label/value rows.
-    /// </summary>
-    public string? Body { get; set; }
-
-    /// <summary>
-    /// Marks this item as a copy-paste code block. Renderers emit a monospace
-    /// &lt;pre&gt; (HTML) / fenced (plain text) / Consolas TextBox with a copy
-    /// button (dialog). Webhooks render only the heading + a "see email or in-app
-    /// dialog for the copy-paste T-SQL" hint when this flag is true.
-    /// </summary>
-    public bool IsCodeBlock { get; set; }
-}
-
-/// <summary>
-/// Serialization DTO for persisting <see cref="AlertContext"/> as JSON.
-/// <see cref="AlertDetailItem.Fields"/> is a <c>List&lt;(string,string)&gt;</c>
-/// ValueTuple, which System.Text.Json will not round-trip (tuple elements are
-/// fields, not properties); these DTOs name every member explicitly so the
-/// persisted context survives the round-trip into the in-app dialog.
-/// <see cref="AlertContext.AttachmentXml"/>/<see cref="AlertContext.AttachmentFileName"/>
-/// are deliberately not persisted (the dialog has no attachment surface).
-/// </summary>
-public record AlertContextDto(List<AlertDetailItemDto> Details);
-public record AlertDetailItemDto(string Heading, List<FieldDto> Fields, string? Body, bool IsCodeBlock);
-public record FieldDto(string Label, string Value);
-
-/// <summary>
-/// Maps <see cref="AlertContext"/> to/from the <see cref="AlertContextDto"/> JSON projection
-/// persisted alongside the flat detail_text. Centralizes the DTO mapping so the persistence
-/// write (EmailAlertService) and the dialog read (AlertDetailWindow) cannot drift.
-/// </summary>
-public static class AlertContextSerializer
-{
-    public static string Serialize(AlertContext context)
-    {
-        var dto = new AlertContextDto(
-            context.Details.ConvertAll(d => new AlertDetailItemDto(
-                d.Heading,
-                d.Fields.ConvertAll(f => new FieldDto(f.Label, f.Value)),
-                d.Body,
-                d.IsCodeBlock)));
-        return JsonSerializer.Serialize(dto);
-    }
-
-    public static bool TryDeserialize(string? json, out AlertContext context)
-    {
-        context = new AlertContext();
-        if (string.IsNullOrWhiteSpace(json))
-            return false;
-
-        try
-        {
-            var dto = JsonSerializer.Deserialize<AlertContextDto>(json);
-            if (dto?.Details is null)
-                return false;
-
-            foreach (var d in dto.Details)
-            {
-                var item = new AlertDetailItem
-                {
-                    Heading = d.Heading,
-                    Body = d.Body,
-                    IsCodeBlock = d.IsCodeBlock
-                };
-                if (d.Fields is not null)
-                {
-                    foreach (var f in d.Fields)
-                        item.Fields.Add((f.Label, f.Value));
-                }
-                context.Details.Add(item);
-            }
-            return true;
-        }
-        catch
-        {
-            return false;
-        }
-    }
-}
